% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FracWorking.R
\name{FracWorking}
\alias{FracWorking}
\title{Fraction of retail workers working on the sample day}
\usage{
FracWorking()
}
\value{
data frame
}
\description{
Fraction of retail workers working on the same day
}

