% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FracWorkingPlot.R
\name{FracWorkingPlot}
\alias{FracWorkingPlot}
\title{Fraction of retail workers working on the sample day, over time}
\usage{
FracWorkingPlot()
}
\value{
data frame
}
\description{
Fraction of retail workers working on the same day, over time
}

