% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HoursReportingPlot.R
\name{HoursReportingPlot}
\alias{HoursReportingPlot}
\title{Plot showing fraction of workers selecting various hours-reports}
\usage{
HoursReportingPlot()
}
\value{
ggplot2
}
\description{
Plot showing fraction of workers selecting various hours-reports
}

