% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ShiftVisualization.R
\name{ShiftVisualization}
\alias{ShiftVisualization}
\title{Visualize the shift patterns for a given year}
\usage{
ShiftVisualization(visualization.year)
}
\arguments{
\item{visualization.year}{}
}
\value{
ggplot2
}
\description{
Visualize the shift patterns for a given year
}

