% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summaryStats.R
\name{summaryStats}
\alias{summaryStats}
\title{Return a dataframe of summary stats by year, conditional upon working}
\usage{
summaryStats()
}
\value{
data.frame
}
\description{
Return a dataframe of summary stats by year
}

